services/vision-ai-enhanced/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
werkzeug==3.0.1
numpy<2.0.0
pillow==10.1.0

# Enhanced Vision and AI dependencies
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0
accelerate==0.25.0
bitsandbytes==0.42.0
sentence-transformers==2.2.2

# Qwen2-VL specific dependencies
qwen-vl-utils==0.0.3
tiktoken==0.5.2

# For multimodal models and API integrations
openai==1.6.0
anthropic==0.8.0
google-generativeai==0.3.2

# gRPC dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Image processing and OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
easyocr==1.7.0

# Performance and monitoring
psutil==5.9.6
prometheus-client==0.19.0

# Configuration and utilities
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Async support
aiohttp==3.9.1
asyncio-mqtt==0.13.0

# Model optimization
optimum==1.16.0
auto-gptq==0.6.0
